+++ app/admin/maintenance_tickets.rb
@@ -62,7 +62,11 @@ ActiveAdmin.register MaintenanceTicket do
               collection: resource.class::STATES,
               include_blank: false
 
-      f.input :assigned_to
+      f.input :assigned_to,
+              as: :select,
+              collection: resource.class::TECH_PEOPLE,
+              include_blank: false
+
       f.input :duration
       f.input :recipients, hint: "Liste d'emails séparés par une virgule"
     end